+++ updateDependencies.sh
@@ -1,13 +1,14 @@
 #!/usr/bin/env bash
-npm install jquery perfect-scrollbar mustache webextension-polyfill underscore i18next i18next-xhr-backend
+npm update jquery perfect-scrollbar mustache webextension-polyfill i18next i18next-xhr-backend
+npm update lodash-cli -g
 
 cssLib='./webextension/data/css/lib'
 jsLib='./webextension/data/js/lib'
 
 
 echo Copying jQuery...
-cp ./node_modules/jquery/dist/jquery.min.js $jsLib
-cp ./node_modules/jquery/dist/jquery.min.map $jsLib
+cp ./node_modules/jquery/dist/jquery.slim.min.js $jsLib
+cp ./node_modules/jquery/dist/jquery.slim.min.map $jsLib
 
 echo Copying mustache...
 cp ./node_modules/mustache/mustache.min.js $jsLib
@@ -20,12 +21,13 @@ echo Copying webextension-polyfill...
 cp ./node_modules/webextension-polyfill/dist/browser-polyfill.min.js $jsLib
 cp ./node_modules/webextension-polyfill/dist/browser-polyfill.min.js.map $jsLib
 
-echo Copying underscore
-cp ./node_modules/underscore/underscore-min.js $jsLib
-cp ./node_modules/underscore/underscore-min.map $jsLib
+echo Copying/Building Lodash Debounce - Custom Build # https://lodash.com/custom-builds
+lodash exports=global include=debounce --production --source-map
+mv lodash.custom.min.js $jsLib
+mv lodash.custom.min.map $jsLib
 
 echo Copying i18next...
-cp ./node_modules/i18next/i18next.min.js $jsLib/../locales
+cp ./node_modules/i18next/i18next.min.js $jsLib
 
 echo Copying i18next-xhr-backend...
-cp ./node_modules/i18next-xhr-backend/i18nextXHRBackend.min.js $jsLib/../locales
+cp ./node_modules/i18next-xhr-backend/i18nextXHRBackend.min.js $jsLib