+++ CHANGELOG.md
@@ -8,6 +8,12 @@
   * Sync support ([Bug 1311710](https://bugzilla.mozilla.org/show_bug.cgi?id=1311710) and ~~[Bug 1220494 (RESOLVED FIXED)](https://bugzilla.mozilla.org/show_bug.cgi?id=1220494)~~ )
   * Websites CSP applied to Content Scripts, and should not happen ([Bug 1267027](https://bugzilla.mozilla.org/show_bug.cgi?id=1267027))
 
+# 11.2.0
+* i: Changed initial load system, with somes files and folders moved
+* i: Moving from Underscore to Lodash, using a custom build to only have debounce
+* i: Moving to the slim version of jQuery
+* Fix: Vocal notification
+
 # 11.1.0
 * +: Stopped using native (Web Extension API) translations to get pural support, with i18next
 * Fix: Panel css